Fix: main_svc And one more correct in sql schema

+++ ewm-main-service/src/main/java/ru/practicum/main/dto/LocationDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 @NoArgsConstructor
 public class LocationDto {
 
-    private double lat;
+    private float lat;
 
-    private double lon;
+    private float lon;
 }
